fix: Better interfaces
