+++ src/components/Card/AddButton.tsx
@@ -1,12 +1,12 @@
 import { useContext } from "react";
-import { CardProps } from ".";
 import Spinner from "../../icons/Spinner";
 import { Context } from "../../data/Context";
 import styles from "./card.module.css";
+import { Item } from "../../data/interfaces";
 
 export interface AddButtonProps {
   shouldBlockButton: boolean;
-  item: CardProps; //HACK: Not really
+  item: Item;
 }
 
 export function AddButton({ shouldBlockButton, item }: AddButtonProps) {
