+++ src/data/interfaces.ts
@@ -1,0 +1,13 @@
+export interface Item {
+  id: number;
+  title: string;
+  description: string;
+  image: string;
+  price: number;
+}
+
+export interface Cart {
+  id: number;
+  quantity: number;
+  productItem: Item;
+}
